+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Fixed
 
+- Avoid refreshing inactive conditional Gemfile dependencies during release lockfile reconciliation.
+
 ### Security
 
 ## [1.2.63] - 2026-08-28